task 1

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -14,14 +14,17 @@ int main(void)
 		/* your code goes there */
 		int b= n % 10;
 		int c= b * b;
-		int a= sqrtf(c);
-		if ( a > 5){
+		int a= sqrt(c);
+		if (a > 5)
+		{
 			printf("Last digit of %d is %d and is greater than 5\n", n, a);
 		}
-		else if( a == 0){
+		else if(a == 0)
+		{
 			printf("Last digit of %d is %d and is 0\n", n, a); 
 		}
-		else if(a < 6 && a != 0) {
+		else if(a < 6 && a != 0)
+		{
 			printf("Last digit of %d is %d and is less than 6 and not 0\n", n, a);
 		}
 		return (0);
